feat: add user table and authentication

+++ backend/src/repositories/task.repository.js
@@ -1,8 +1,6 @@
 import pool from "../db/pool.js";
 
-
 export async function findAllTasks() {
-
   const result = await pool.query(`
     SELECT
       id,
@@ -10,6 +8,7 @@ export async function findAllTasks() {
       description,
       due_date,
       completed,
+      priority,
       created_at,
       updated_at
     FROM tasks
@@ -19,57 +18,49 @@ export async function findAllTasks() {
   return result.rows;
 }
 
-
-export async function findTaskById(id) {
-
+export async function findTaskById(id, userId) {
   const result = await pool.query(
     `
       SELECT
+        user_id,
         id,
         title,
         description,
         due_date,
         completed,
+        priority,
         created_at,
         updated_at
       FROM tasks
       WHERE id = $1
+        AND user_id = $2
     `,
-    [id]
+    [id, userId],
   );
 
   return result.rows[0] ?? null;
 }
 
-
 export async function createTask({
+  userId,
   title,
   description,
-  dueDate
+  dueDate,
+  priority,
 }) {
-
   const result = await pool.query(
     `
       INSERT INTO tasks (
-        title,
-        description,
-        due_date
-      )
-      VALUES ($1, $2, $3)
-      RETURNING
-        id,
+        user_id,
         title,
         description,
         due_date,
-        completed,
-        created_at,
-        updated_at
+        priority
+      )
+      VALUES ($1, $2, $3, $4, $5)
+      RETURNING *
     `,
-    [
-      title,
-      description ?? null,
-      dueDate ?? null
-    ]
+    [userId, title, description ?? null, dueDate ?? null, priority],
   );
 
   return result.rows[0];
@@ -77,9 +68,8 @@ export async function createTask({
 
 export async function updateTask(
   id,
-  { title, description, dueDate, completed }
+  { title, description, dueDate, completed },
 ) {
-
   const result = await pool.query(
     `
       UPDATE tasks
@@ -104,22 +94,22 @@ export async function updateTask(
       description ?? null,
       dueDate ?? null,
       completed ?? null,
-      id
-    ]
+      id,
+    ],
   );
 
   return result.rows[0] ?? null;
 }
 
-export async function deleteTask(id) {
-
+export async function deleteTask(id, userId) {
   const result = await pool.query(
     `
       DELETE FROM tasks
-      WHERE id = $1
+      WHERE id = $1 
+      AND user_id = $2
       RETURNING id
     `,
-    [id]
+    [id, userId],
   );
 
   return result.rows[0] ?? null;
